+++ CHANGELOG.md
@@ -36,7 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * New method signature for `isolines()` having an intervals parameter that specifies the number of contour levels to generate.
 * New method signature for `straightSkeleton()` that accepts an integer to control the number of nearest neighboring edges considered during collision detection.
 * `contrastField()` to `PGS_Contour`. Generates vector contour lines representing a "contrast field" of a shape with respect to a given reference point.
-
+* `arcDivision` to `PGS_Tiling`. Creates a cellular partition of the plane using arcs formed by circles seeded along its boundary.
+* `sliceDivision` to `PGS_Tiling`. Divides the plane into randomly “sliced” polygonal regions.
 
 ### Changes
 * Optimised `PGS_CirclePacking.tangencyPack()`. It's now around 1.5-2x faster and has higher precision.